+++ integration/src/harness.rs
@@ -1,9 +1,9 @@
 use std::{
+    path::PathBuf,
     sync::{
         Arc, atomic::{AtomicUsize, Ordering}, mpsc, mpsc::channel, Mutex,
         RwLock,
     },
-    path::PathBuf,
     thread, time,
 };
 
@@ -144,7 +144,7 @@ pub fn new_server(port: u16) -> Server {
 
         // Register a getter for the variable
         if let Some(ref mut v) = address_space.find_variable_mut(v1_node.clone()) {
-            let getter = AttrFnGetter::new(move |_, _, _| -> Result<Option<DataValue>, StatusCode> {
+            let getter = AttrFnGetter::new(move |_, _, _, _, _| -> Result<Option<DataValue>, StatusCode> {
                 Ok(Some(DataValue::new(100)))
             });
             v.set_value_getter(Arc::new(Mutex::new(getter)));