+++ server/src/address_space/base.rs
@@ -3,7 +3,7 @@ use opcua_types::{
     status_code::StatusCode,
 };
 
-use super::node::{NodeBase, Node};
+use super::node::{Node, NodeBase};
 
 /// Base node class contains the attributes that all other kinds of nodes need. Part 3, diagram B.4
 #[derive(Debug)]
@@ -71,7 +71,7 @@ impl NodeBase for Base {
 }
 
 impl Node for Base {
-    fn get_attribute_max_age(&self, attribute_id: AttributeId, _max_age: f64) -> Option<DataValue> {
+    fn get_attribute_max_age(&self, attribute_id: AttributeId, _index_range: NumericRange, _data_encoding: &QualifiedName, _max_age: f64) -> Option<DataValue> {
         match attribute_id {
             AttributeId::NodeClass => Some(DataValue::new(self.node_class as i32)),
             AttributeId::NodeId => Some(DataValue::new(self.node_id())),