+++ server/src/address_space/method.rs
@@ -82,11 +82,11 @@ impl Default for Method {
 node_base_impl!(Method);
 
 impl Node for Method {
-    fn get_attribute_max_age(&self, attribute_id: AttributeId, max_age: f64) -> Option<DataValue> {
+    fn get_attribute_max_age(&self, attribute_id: AttributeId, index_range: NumericRange, data_encoding: &QualifiedName, max_age: f64) -> Option<DataValue> {
         match attribute_id {
             AttributeId::Executable => Some(Variant::from(self.executable()).into()),
             AttributeId::UserExecutable => Some(Variant::from(self.user_executable()).into()),
-            _ => self.base.get_attribute_max_age(attribute_id, max_age)
+            _ => self.base.get_attribute_max_age(attribute_id, index_range, data_encoding, max_age)
         }
     }