+++ server/src/address_space/reference_type.rs
@@ -2,7 +2,7 @@
 
 use opcua_types::service_types::ReferenceTypeAttributes;
 
-use crate::address_space::{base::Base, node::NodeBase, node::Node};
+use crate::address_space::{base::Base, node::Node, node::NodeBase};
 
 node_builder_impl!(ReferenceTypeBuilder, ReferenceType);
 node_builder_impl_subtype!(ReferenceTypeBuilder);
@@ -30,12 +30,12 @@ impl Default for ReferenceType {
 node_base_impl!(ReferenceType);
 
 impl Node for ReferenceType {
-    fn get_attribute_max_age(&self, attribute_id: AttributeId, max_age: f64) -> Option<DataValue> {
+    fn get_attribute_max_age(&self, attribute_id: AttributeId, index_range: NumericRange, data_encoding: &QualifiedName, max_age: f64) -> Option<DataValue> {
         match attribute_id {
             AttributeId::Symmetric => Some(Variant::from(self.symmetric()).into()),
             AttributeId::IsAbstract => Some(Variant::from(self.is_abstract()).into()),
             AttributeId::InverseName => self.inverse_name().map(|v| Variant::from(v).into()),
-            _ => self.base.get_attribute_max_age(attribute_id, max_age)
+            _ => self.base.get_attribute_max_age(attribute_id, index_range, data_encoding, max_age)
         }
     }