+++ server/src/events/operator.rs
@@ -5,10 +5,10 @@ use std::convert::TryFrom;
 use regex::Regex;
 
 use opcua_types::{
-    AttributeId, ExtensionObject, Variant, VariantTypeId, NodeId,
+    AttributeId, ExtensionObject, NodeId, NumericRange, operand::Operand, QualifiedName, service_types::{ContentFilterElement, FilterOperator, SimpleAttributeOperand},
     status_code::StatusCode,
-    operand::Operand,
-    service_types::{ContentFilterElement, FilterOperator, SimpleAttributeOperand},
+    Variant,
+    VariantTypeId,
 };
 
 use crate::address_space::{
@@ -98,7 +98,7 @@ pub(crate) fn value_of_simple_attribute(object_id: &NodeId, o: &SimpleAttributeO
                 }
                 NodeType::Variable(ref node) => {
                     if o.attribute_id == AttributeId::Value as u32 {
-                        if let Some(ref value) = node.value().value {
+                        if let Some(ref value) = node.value(NumericRange::None, &QualifiedName::null()).value {
                             value.clone()
                         } else {
                             Variant::Empty