+++ server/src/tests/address_space.rs
@@ -432,7 +432,7 @@ fn variable_builder() {
     assert_eq!(v.value_rank(), 10);
     assert_eq!(v.array_dimensions().unwrap(), vec![1, 2, 3]);
     assert_eq!(v.historizing(), true);
-    assert_eq!(v.value().value.unwrap(), Variant::from(999));
+    assert_eq!(v.value(NumericRange::None, &QualifiedName::null()).value.unwrap(), Variant::from(999));
     assert_eq!(v.minimum_sampling_interval().unwrap(), 123.0);
 
     // Add a variable to the address space
@@ -488,7 +488,7 @@ fn method_builder() {
         // verify OutputArguments / Argument value
         assert_eq!(v.data_type(), DataTypeId::Argument.into());
         assert_eq!(v.display_name(), LocalizedText::from("OutputArguments"));
-        let v = v.value().value.unwrap();
+        let v = v.value(NumericRange::None, &QualifiedName::null()).value.unwrap();
         if let Variant::Array(v) = v {
             assert_eq!(v.len(), 1);
             let v = v.get(0).unwrap().clone();